Introduce InstrumentationTimingVerificationTests
Allows us to make a more specific abstract class for the existing TCKs while allowing for future/third-party TCKs that may not be based on timing instrumentation.
