+++ micrometer-test/src/main/java/io/micrometer/core/instrument/InstrumentationVerificationTests.java
@@ -30,9 +30,7 @@ import org.junit.platform.commons.util.AnnotationUtils;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Parameter;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 abstract class InstrumentationVerificationTests {
 
@@ -75,7 +73,7 @@ abstract class InstrumentationVerificationTests {
         METRICS_VIA_METER_REGISTRY,
 
         /**
-         * Runs the tests by using the Observation API.
+         * Runs the tests by using the Observation API and a MeterObservationHandler.
          */
         METRICS_VIA_OBSERVATIONS_WITH_METRICS_HANDLER
